requirements.txt
pandas==2.0.2
numpy==1.24.3
nltk==3.8.1
scikit-learn==1.2.2
tqdm==4.65.0
gensim==4.3.1
pytorch==2.0.1
matplotlib==3.7.1